Add dockerfile and shellcheck linters, refactore hack dir
