+++ Makefile
@@ -8,7 +8,10 @@ export GOPROXY = https://proxy.golang.org
 all: build-race test-unit test-integration test-lint
 .PHONY: all
 
-# Build
+############
+# Building #
+############
+
 build:
 	go build -o codeowners-validator ./main.go
 .PHONY: build
@@ -17,23 +20,34 @@ build-race:
 	go build -race -o codeowners-validator ./main.go
 .PHONY: build-race
 
-# Test
+###########
+# Testing #
+###########
+
 test-unit:
-	./hack/ci/run-test-unit.sh
+	./hack/run-test-unit.sh
 .PHONY: test-unit
 
 test-integration: build
-	env BINARY_PATH=$(PWD)/codeowners-validator ./hack/ci/run-test-integration.sh
+	env BINARY_PATH=$(PWD)/codeowners-validator ./hack/run-test-integration.sh
 .PHONY: test-integration
 
 test-lint:
-	./hack/ci/run-lint.sh
+	./hack/run-lint.sh
 .PHONY: test-lint
 
 test-hammer:
 	go test -count=100 ./...
 .PHONY: test-hammer
 
-cover-html: test-unit
+test-unit-cover-html: test-unit
 	go tool cover -html=./coverage.txt
 .PHONY: cover-html
+
+###############
+# Development #
+###############
+
+fix-lint-issues:
+	LINT_FORCE_FIX=true ./hack/run-lint.sh
+.PHONY: fix-lint
