+++ hack/README.md
@@ -1,0 +1,7 @@
+# Hack directory
+
+This package contains various scripts that are used by Codeowners Validator developers.
+
+## Purpose
+
+This directory contains tools, such as Go fmt, Go lint, and Go vet, that help to maintain the source code compliant to Go best coding practices. It also includes utility scripts that generate code, and scripts executed on CI pipelines.
