+++ hack/lib/utilities.sh
