+++ internal/check/duplicated_pattern_test.go
@@ -61,5 +61,4 @@ func TestDuplicatedPattern(t *testing.T) {
 			assert.ElementsMatch(t, tc.expectedIssues, out.Issues)
 		})
 	}
-
 }
