+++ internal/check/not_owned_file.go
@@ -7,7 +7,7 @@ import (
 	"path"
 	"strings"
 
-	ctxutil "github.com/mszostok/codeowners-validator/internal/context"
+	"github.com/mszostok/codeowners-validator/internal/ctxutil"
 	"github.com/mszostok/codeowners-validator/pkg/codeowners"
 
 	"github.com/hashicorp/go-multierror"
