+++ internal/ctxutil/check.go
@@ -1,4 +1,4 @@
-package context
+package ctxutil
 
 import "context"
